+++ test/unit/sonarqube_code_coverage.sh
@@ -13,6 +13,7 @@ coverage run -a --source=mongo_db_data test/unit/mongo_db_data/help_message.py
 coverage run -a --source=mongo_db_data test/unit/mongo_db_data/insert_doc.py
 coverage run -a --source=mongo_db_data test/unit/mongo_db_data/insert_doc2.py
 coverage run -a --source=mongo_db_data test/unit/mongo_db_data/main.py
+coverage run -a --source=mongo_db_data test/unit/mongo_db_data/post_process.py
 coverage run -a --source=mongo_db_data test/unit/mongo_db_data/process_args.py
 coverage run -a --source=mongo_db_data test/unit/mongo_db_data/run_program.py
 coverage run -a --source=mongo_db_data test/unit/mongo_db_data/truncate_coll.py